+++ packages/cds-annotation-parser/src/transformer/ast.ts
@@ -1,5 +1,5 @@
 import type { CstNode, IToken, CstNodeLocation } from 'chevrotain';
-import { Range, Position } from 'vscode-languageserver-types';
+import { Range, Position } from '@sap-ux/text-document-utils';
 
 import type {
     Annotation,