+++ packages/fe-fpm-writer/src/building-block/prompts/questions/index.ts
@@ -1,0 +1,4 @@
+export * from './building-blocks';
+export * from './chart';
+export * from './filter-bar';
+export * from './table';