+++ examples/odata-cli/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @sap-ux/odata-cli
 
+## 0.15.3
+
+### Patch Changes
+
+-   Updated dependencies [0084205]
+    -   @sap-ux/axios-extension@1.16.3
+
+## 0.15.2
+
+### Patch Changes
+
+-   e69db46: Upgrade fast-xml-parser
+-   Updated dependencies [e69db46]
+    -   @sap-ux/axios-extension@1.16.2
+
 ## 0.15.1
 
 ### Patch Changes