+++ packages/ui5-library-writer/CHANGELOG.md
@@ -1,5 +1,61 @@
 # @sap-ux/ui5-library-writer
 
+## 0.5.11
+
+### Patch Changes
+
+-   Updated dependencies [cc16cbb]
+    -   @sap-ux/project-access@1.26.5
+    -   @sap-ux/ui5-application-writer@1.1.3
+
+## 0.5.10
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.26.4
+-   @sap-ux/ui5-application-writer@1.1.3
+
+## 0.5.9
+
+### Patch Changes
+
+-   Updated dependencies [88c8bf6]
+    -   @sap-ux/project-access@1.26.3
+    -   @sap-ux/ui5-application-writer@1.1.3
+
+## 0.5.8
+
+### Patch Changes
+
+-   e69db46: Upgrade fast-xml-parser
+-   Updated dependencies [e69db46]
+    -   @sap-ux/project-access@1.26.2
+    -   @sap-ux/ui5-application-writer@1.1.3
+
+## 0.5.7
+
+### Patch Changes
+
+-   Updated dependencies [a986655]
+    -   @sap-ux/project-access@1.26.1
+    -   @sap-ux/ui5-application-writer@1.1.2
+
+## 0.5.6
+
+### Patch Changes
+
+-   Updated dependencies [518bf7e]
+    -   @sap-ux/project-access@1.26.0
+    -   @sap-ux/ui5-application-writer@1.1.2
+
+## 0.5.5
+
+### Patch Changes
+
+-   Updated dependencies [99b7b5f]
+    -   @sap-ux/project-access@1.25.8
+    -   @sap-ux/ui5-application-writer@1.1.2
+
 ## 0.5.4
 
 ### Patch Changes